requirements.txt
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
decorator==5.0.9
Django==2.2.18
django-cors-headers==3.7.0
django-extensions==3.1.1
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
